Moved replace_parameters to _utils/helpers.py & madeit member of _utils
