+++ nilearn/_utils/__init__.py
@@ -8,7 +8,9 @@ from .numpy_conversions import as_ndarray
 from .cache_mixin import CacheMixin
 
 from .logger import _compose_err_msg
+from nilearn._utils.helpers import replace_parameters
 
 __all__ = ['check_niimg', 'check_niimg_3d', 'concat_niimgs', 'check_niimg_4d',
            '_repr_niimgs', 'copy_img', 'load_niimg',
-           'as_ndarray', 'CacheMixin', '_compose_err_msg']
+           'as_ndarray', 'CacheMixin', '_compose_err_msg', 'replace_parameters',
+           ]
